feat: add search benchmarks
