+++ desktop/Cargo.toml
@@ -23,3 +23,8 @@ lru = "0.12"
 
 [dev-dependencies]
 tempfile = "3"
+criterion = "0.5"
+
+[[bench]]
+name = "search"
+harness = false
